# tests/contract/api-gateway-openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectorListItemResponse": {
        "description": "Response model for connector list items (combines definition and config).",
        "properties": {
          "auth_type": {
            "title": "Auth Type",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "config_fields": {
            "items": {
              "type": "object"
            },
            "title": "Config Fields",
            "type": "array"
          },
          "config_schema": {
            "items": {
              "type": "object"
            },
            "title": "Config Schema",
            "type": "array"
          },
          "configured": {
            "default": false,
            "title": "Configured",
            "type": "boolean"
          },
          "connector_id": {
            "title": "Connector Id",
            "type": "string"
          },
          "custom_fields": {
            "title": "Custom Fields",
            "type": "object"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "env_vars": {
            "items": {
              "type": "string"
            },
            "title": "Env Vars",
            "type": "array"
          },
          "health_status": {
            "default": "unknown",
            "title": "Health Status",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "instance_url": {
            "default": "",
            "title": "Instance Url",
            "type": "string"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_key": {
            "default": "",
            "title": "Project Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "supported_sync_directions": {
            "items": {
              "type": "string"
            },
            "title": "Supported Sync Directions",
            "type": "array"
          },
          "sync_direction": {
            "default": "inbound",
            "title": "Sync Direction",
            "type": "string"
          },
          "sync_frequency": {
            "default": "daily",
            "title": "Sync Frequency",
            "type": "string"
          }
        },
        "required": [
          "connector_id",
          "name",
          "description",
          "category",
          "status",
          "icon",
          "supported_sync_directions",
          "auth_type",
          "config_fields",
          "config_schema",
          "env_vars"
        ],
        "title": "ConnectorListItemResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-native Project Portfolio Management platform with 25 specialized agents",
    "title": "Multi-Agent PPM Platform",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/connectors": {
      "get": {
        "description": "List all available connectors with their configuration status.\n\nOptionally filter by category.",
        "operationId": "list_connectors_api_v1_connectors_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectorListItemResponse"
                  },
                  "title": "Response List Connectors Api V1 Connectors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Connectors",
        "tags": [
          "connectors"
        ]
      }
    },
...
    "/v1/connectors/{connector_id}": {
      "get": {
        "description": "Get a specific connector with its configuration.",
        "operationId": "get_connector_api_v1_connectors__connector_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorListItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Connector",
        "tags": [
          "connectors"
        ]
      }
    },
...
  }
}